Began adapting TalkNexus UI and ollama interfaces to support chaperone overlay
