+++ .gitignore
@@ -1,4 +1,5 @@
-
+.venv
+.vscode
 /st_pages/__pycache__
 /schema/__pycache__
 /saved_conversations
