+++ src/navigation.py
@@ -2,10 +2,13 @@ import warnings
 warnings.filterwarnings("ignore", category=Warning)
 
 import streamlit as st
-import st_pages # required modules
+import imgs
+import st_pages.home
+import st_pages.ai_chatbot
+import st_pages.model_management # required modules
 
 # Set page config
-st.set_page_config(page_title="TalkNexus - Ollama Chatbot Multi-Model Interface", layout="wide", page_icon="🤖")
+st.set_page_config(page_title="Anti-Parasocial Chaperones", layout="wide", page_icon="🤖")
 
 # Load custom CSS from file
 def load_css(file_name):
@@ -16,35 +19,23 @@ load_css('styles.css')
 
 # Initialize session state
 if 'current_page' not in st.session_state:
-    st.session_state.current_page = "Home"
+    st.session_state.current_page = "AI Conversation"
 
 # Header
 st.markdown(f"""
 <div class="header">
     <div class="animated-bg"></div>
     <div class="header-content">
-        <h1 class="header-title">Ollama Chatbot Multi-Model Interface</h1> 
-        <p class="header-subtitle">Advanced Language Models & Intelligent Conversations</p>
+        <h1 class="header-title">Anti-Parasocial Chaperones</h1> 
+        <p class="header-subtitle">A side-by-side comparison</p>
     </div>
 </div>
 """, unsafe_allow_html=True)
 
 # Enhanced pages definition
 PAGES = {
-    "Home": {
-        "icon": "house-door",
-        "func": st_pages.home,
-        "description": "Guidelines & Overview",
-        "badge": "Informative",
-        "color": "var(--primary-color)"
-    },
-    "Language Models Management": {
-        "icon": "gear",
-        "func": st_pages.model_management,
-        "description": "Download Models",
-        "badge": "Configurations",
-        "color": "var(--secondary-color)"
-    },
+    
+   
     "AI Conversation": {
         "icon": "chat-dots",
         "func": st_pages.ai_chatbot,
@@ -52,13 +43,7 @@ PAGES = {
         "badge": "Application",
         "color": "var(--highlight-color)"
     },
-    "RAG Conversation": {
-        "icon": "chat-dots",
-        "func": st_pages.rag_chat,
-        "description": "PDF AI Chat Assistant",
-        "badge": "Application",
-        "color": "var(--highlight-color)"
-    }
+  
 }
 
 st.markdown("""
@@ -73,7 +58,7 @@ def navigate():
                 <div class="profile-section">
                     <div class="profile-info">
                         <h1 style="font-size: 38px;">TalkNexus</h1>
-                        <span class="active-badge" style="font-size: 20px;">AI Chatbot Multi-Model Application</span>
+                        <span class="active-badge" style="font-size: 20px;"></span>
                     </div>
                 </div>
             </div>
@@ -127,16 +112,9 @@ try:
     page_function = PAGES[selected_page]["func"]
     page_function()
 except Exception as e:
-    st.error(f"Error loading page: {str(e)}")
-    st_pages.home.run()
+    pass
+    #st.error(f"Error loading page: {str(e)}")
 
 # Display the footer
 st.markdown("""
-<div class="footer">
-    <div class="footer-content">
-        <p>© 2024 Powered by <a href="https://github.com/TsLu1s" target="_blank">TsLu1s </a>. 
-        Advanced Language Models & Intelligent Conversations
-        | Project Source: <a href="https://github.com/TsLu1s/talknexus" target="_blank"> TalkNexus</p>
-    </div>
-</div>
 """, unsafe_allow_html=True)
